1.message接口新增方法，原方法已过时 2.order模块修改 3.股票股指相关重构 by 杨日东

+++ financial_management/src/main/java/financial_management/bl/product/InvestmentService.java
@@ -17,6 +17,6 @@ public interface InvestmentService {
 
     List<InvestRecProductVO> getAllInvProduct();
 
-    BasicResponse purchase(Long userId, String code, Integer amount, Double totalprice, String type);
+    BasicResponse purchase(Long userId, String code, Float amount, Double totalprice, String type);
 
 }
